+++ inst/ui-server/server.R
@@ -1,0 +1,3 @@
+library(auth0)
+
+auth0_server(function(input, output, session) {})